Modify max-logs default to 20, clarify docs
This change increases the default to 20. The code checks if max-logs was decreased from
the previous value, and aborts. The src/misc.c hard-coded default is purposefully
mismatched to the config file default setting, as the misc.c value is first read at load
prior to being modified by the config setting. If the user were to set a value of less
than the default misc.c value, eggdrop detects this as a 'decrease' in value and would
not start. Thus, 5 is used as the hard-code default.