Show unset entries in getuser handle as empty strings.
Settings that weren't initialized on a user are now shown as the empty string in Tcl.
The reason is that it could be relevant to know which settings aren't set.
There is no way to unset them other than to set them to the empty strings,
so this is consistent behavior. One might also want to know unset/empty settings,
to prompt user to input them or fill in sane defaults.