Check for compatibility OpenSSL functions regardless of whether paths to the library were specified.
The functions ASN1_STRING_get0_data, hex_to_string, string_to_hex need to be checked regardless of
whether the user specified the --with-ssllib= switch. Previously, it was only checked if they were not
specified. Which is suitable in finding the library, but we need those compatibility macros in
config.h regardless of that. ASN1_STRING_get0_data has a compatibilty version in eggdrop, the
other two can result in an unsuitable SSL library and fail the check. Fixes #425.

Found by: Artea, Zela