Connect to IP of the already-active link for userfile xfer. Fixes #21
Found by: Darkreap1, simple
Patch by: Cizzle

A leaf will now use the address of the hub (to which it's already successfully connected) instead of the IP given by the hub which was wrong in certain cases (especially in NAT situations). Backward compatibility is kept, older bots still get the IP given by the hub.