Fix shared userfile missing first 511 bytes when using SSL
Found by: Geo
Patch by: Cizzle

Fixes sharing of a userfile with SSL which would stall with 511 bytes missing (as could be seen in .dccstat output) and eventually fail. Racing condition where the ssl handshake was finished between the check and reading from the socket, leaving a flag active that treated the data as garbage.