Change addbot command to use spaces. (#496)
Use spaces in the address instead of '/' or ':'. Command is now addbot handle address [bot-port [relay-port]].
For backward compatibility, address can still be built up like host[:bot-port[/relay-port]]. If host is in IPv6 format, ports van only be specified when using '[' ']'.
The separate port arguments are ignored if the old address format is used.