Fix order raw/builtin binds are called. Closes #354 and #387
Found by: jack3 and maimizuno
Patch by: michaelortmann
Fixes: #354 and #387

Sometimes the raw bind is executed before the builtin invite, sometimes the other way around.
The builtin invite modifies the args in chan.c gotinvite()

The mysterious "toggle" was found, probably an optimization, moving last recently used binds to the front of the bind table. This leads to toggling the order of execution every time the same binds are triggered. Now the optimization is only done when there is exactly 1 bind for the trigger.
No more toggle, no more truncation/substitution.