Enhance botnet memory management
 Patch by: michaelortmann

Botnet parties are stored in an array of party_t. Eggdrop started with array size 50 (even for single/unlinked bots, this wastes 4K of memory) and if it gets full adds 50 more. With this patch, eggdrop will start with size 1 and if it gets full doubles size. This is a good algorithm we already use for memtbl. Additionally this patch also makes use of the known lengths of the tand bot list, so we dont need to iterate there.