Convert net-type to string
Patch by: michaelortmann, Geo

This patch removes references to the old 0-5 settings of net-type and replaces them with string vales, but still accepts the integers for backwards-compatibility reasons. Additionally, this patch updates the configured network settings with current ircd values, as well as adds network settings for Rizon, Freenode, and Quakenet.

Note to the next guy/gal who looks to update these settings: For most ircds now, max-modes is a summation of the individual modes. ie, in 005 you see MAXLIST=beI:100; that means the total combination of b+e+I can not exceed 100 (MAXLIST in this case sums all the modes set to get the maximum modes that can be set for a channel). If you see MAXLIST=b: 200,e:100,I:100, that means the total combination is still a summation of each of those modes, but each mode has its own limit. Thus, maxmodes is 400 (200+100+100), slightly different then the previous example.

* update default nick-len by net-type and enhance net-types - last station before 005
* make net-type a string
* Bump server.mod version