Restrict b/e/I to <63 char hostmask
Found by: @michaelortmann
Patch by: Geo
Fixes: #887

1035/2812 states:
<hostname> has a maximum length of 63 characters. This is a limitation of the protocol as internet hostnames (in particular) can be longer. Such restriction is necessary because IRC messages are limited to 512 characters in length. Clients connecting from a host which name is longer than 63 characters are registered using the host (numeric) address instead of the host name.

As previously written, eggdrop was checking a ban mask length in its entirety, to include nick and ident, and then truncating to a 70-char length for the mask as a whole. This patch updates the check to only look at the hostname portion of a mask, not the mask in its entirety, and then truncate appropriately.