Listen for connections on multiple IPs
Closes: #151

Adds an optional IP address argument to the 'listen' command, allowing it to be run multiple times, creating a new listening interface each time. This will deprecate the config setting 'listen-addr' as it is not needed anymore, but the ability to use listen-addr remains in code to allow backwards compatibility with older configs.