Prevent crash on multi-listen rehash Found by: PeGaSuS Patch by: Geo
When rehashing and listening on 0.0.0.0 or , Eggdrop was improperly checking and would quit. More research also showed the lack of accomodation of IPv6 addresses. Finally, fixed logic so we can now listen on the same port on multiple addresses. Also, I pity whoever tries to decipher the nested logic of that last if statement.