Feat/pbkdf2 (#1055)
Found by: Various
Patch by: thommey and michaelortmann
Fixes: #810 #811

Adds PBKDF2 support for hashing. The intent/hope here is that someone will write a proper PASS2 "umbrella" module in the future which will serve as a register for modules such as PBKDF2, and pass values for hashing to the appropriate sub-module based on the PHC string format used for the hash (ie, $pbkdf2-SHA256$rounds=5000$hipwbiJ.... goes to PBKDF2 if loaded, $scrypt-SHA256$rounds=5000$hipwbiJ.... goes to scrypt if loaded, etc).

PASSWORDMAX is now 30. Rounds is defaulted to 16,000, default method is SHA256.

I won't re-hash the docs in this commit message, but it is important to note that this module was written for compatibility to be co-loaded with the blowfish module; it will maintain both the PASS and PASS2 fields. The expectation is that this feature will be removed for Eggdrop 2.0, WHICH IS DEFINITELY GOING TO HAPPEN IN MY LIFETIME. No co-load of blowfish = no maintenance of the PASS field.

If the method used for hashing changes (ie, SHA256 to SHA512), the hashes will be updated to the new method when the user next logs in.

There is no encrypt/decrypt functionality at this time; this is definitely a punt on our part, apologies to the community. Feel free to submit another PR :)