Make core dns async
Found by: michaelortmann
Patch by: michaelortmann
Fixes: lots of dns issues

The idea is to make eggdrops core dns use libc resolver functions in a non-blocking way. This makes the dns mod superflous. This solution uses pthreads, which is available under posix 2001 and cygwin, so we are fine using it now. Back when the dns mod was created it was not such a good option, but now it is. As a result of these changes DNS nameservers are (necessarily) controlled via the host's resolv.conf, not eggdrop.conf

Because of the breadth of change and fact that testing this will take a large amount of time to find issues from all the possible host configurations out there, this is merged in as a 'beta' feature and can only be enabled via the --enable-tdns compile flag. The DNS mod is prevented from loading if this feature is enabled. The plan is to remove the ifdefs to make it 'the way' at a point in time in the future once we have better confidence it is stable, whether that be 1.9.1 or 2.0.