Add granular userfile sharing flags
Patch by: Cizzle
Closes: #1
Closes: #7 

One-line summary: Introduces new bot flags bcejnu to allow a passive bot (slave) of sharing changes to bans, channels, exempts, invites, ignores and users respectively with an aggressive bot (hub). Also introduces bot flag d for when you don't want to share anything, as bot flag s will still allow anything to be shared.

The caution, however - this allows linked bots, sharing userfiles, to become desynchronized. If I add the leafbot on the hub with a bottattr of +n and then try and do a .+user user1 on the leafbot, the leafbot still adds that user to the local userfile. I guess linked bots that don't share userfiles are already going to be in the exact situation so it's not crazy to see this state, but that user would be over-written on the next userfile download from the hub, which doesn't happen on non-sharing linked bots.

And hey, how cool is it to close #1 ?!?!