RAWT blocks RAW
Found by: DasBrain
Patch by: Geo, thommey
Fixes: #1164

Docs state RAW and RAWT binds should block any further processing, but because they were both called, if one returned a 1, the other might not, and the line would still be processed further. This PR gives precedence to the RAWT bind over a RAW bind (in other words, if the RAWT bind returns a 1, the RAW bind will not be processed. If a RAW bind returns a 1, the RAWT still triggers). If a user encounters issues with this, it is suggested to switch their RAW binds to RAWT binds.