Use nick/altnick/randnick on bad nickname
If an eggdrop joins a server and attempts to register with an invalid nick, it looks at keep-nick in the config. If keep-nick was enabled, it would not try another nickname and end up disconnecting. If keep-nick was not enabled, it would generate a random nickname to continue registration with.

This PR modifies the behavior to instead remove a reliance on keep-nick, and first try the altnick, and THEN a random nick, with the thinking that getting the eggdrop online is more important than preserving an invalid nickname, and will also simplify the troubleshooting for a user from 'eggdrop doesnt run' to 'eggdrop has a weird nickname'