Fix reliance on select() emptying sets on return 0, it does not seem to
On a docker instance of eggdrop stdin is a pipe if run with -i.
Eggdrop blocked forever on a read(stdin) because stdin was assumed
readable even though select returns 0.
The only explanation is that select returning 0 does not clear the
input FD sets. It is pointless to loop through all sockets then anyway.

Found by: Geo
Patch by: thommey