Add account tracking to Eggdrop
This merge adds a basic account tracking capability to Eggdrop. Reliable account tracking requires three server components: WHOX responses enabled, the extended-join capability, and the account-notify capability. Without WHOX supported by the server and extended-join and account-notify capabilities negotiated, account tracking is not reliable. To make this check easy for Tcl scripts, the "accounttracking" variable was implemented and can be easily checked to see if account-tracking is currently in a reliable state. Other major changes this PR implements:

- Adds ACCOUNT getuser/setuser setting
- Adds +account/-account commands to command line, to add an account to a handle
- Adds account2hand Tcl command that takes a nickname, checks if an account is associated and, if so, returns the handle the account is added to (if any)