Fix plain SASL when TLS unavailable
Found by: @eryg-kai
Patch by: @eryg-kai

Fix plain SASL auth when TLS is unavailable

There is common code after the if/else that was being omitted by an ifdef that should only be affecting the last else branch. This change brings this ifdef in line with the previous ifdef (utilizing an early `return 1`).