Fixes a DNS bug when DCC connections were disconnected while the DNS request was pending. Fixes #1174 (#1319)
Found by: DasBrain, ERR1R, CrazyCat
Patch by: thommey
Fixes: #1174 #1175

In a host->IP scenario, devent_t contains a malloc'd copy of the hostname.
In the IP->host scenario however, devent_t contained an outside allocated pointer
to ip_addr (sockname_t).
In the Tcl case it was embedded in the tclinfo structure, but in the DCC case
it pointed to the idx information in dcc[i].
"Lost connection" wipes that idx info and invalidates it, invalidating
the information about the pending dns request. That would then get stuck with
invalid data.

As soon as the idx is re-used however, the address is overwritten by answer(),
and dns thinks it already sent a request for that IP, also it gets stuck.

The fix treats IP->host the same as host->IP and dynamically allocates the IP
as it does with the hostname.