Workaround for TCL_UTF_MAX==3
Found by: Empus
Patch by: thommey
Fixes: #1163

Ugly workaround that does not break current usage of emoji in the Tcl API. Our usage of the Tcl API is not correct because Eggdrop is encoding agnostic, and processes byte arrays only. Those however get passed to Tcl as strings, when they are not. The real solution is fairly complex (decode all bytes into strings with the proper encoding guess from/to the Tcl API).

If users are unable to recompile Tcl with increased TCL_UTF_MAX and are stuck with 8.6 stable (8.7 is not released as stable at the time of this PR), this should be a suitable, albeit ugly and complex, workaround that keeps emoji support alive.

We do not wish to break the currently working emoji usage however, which mostly works in the current state, except for a crash if passed to [string tolower/toupper/totitle], other problematic call sites aren't known.

So we overwrite Tcl's string tolower/toupper/totitle with our own version that:

    detects utf-8 sequences that are 4 bytes
    converts them into utf-8 encoded surrogate pairs before Tcl's string functions are called
    those remain untouched by case changes in Tcl
    convert the surrogate pairs back in the result to the original utf-8 4-byte code sequences

Side effect is that we also convert surrogate pairs if they were sent as such, which should not have a negative impact on existing scripts.

Single-byte encodings being valid 4-byte emoji sequences or 6-byte surrogate pairs is hopefully highly unlikely.