Add Autoscript feature
The autoscripts system will allow Eggdrop to download, install, and configure scripts via the partyline. Since it's inception, users have needed to load a Tcl script into Eggdrop by downloading a Tcl file, editing the file to customize settings, and then sourcing that file in the config file. The autoscripts system helps by:

    Centralizing commonly-used scripts in a single location
    Allowing scripts to be downloaded via the partyline
    Allowing script settings to be configured via the partyline
    Allowing user-written scripts to be managed by the autoscripts system
    Providing a documented API to write autoscripts-compatible scripts

To take advantage of this, scripts will have to be modified to adhere to a published API and be accompanied by a manifest.json file. Eggheads will host commonly used scripts so that users can download known "good" scripts. Additionally, the API will allow scripts to be packaged and distributed outside of being downloaded from eggheads.

An update function is still To-Do