Add got-chanlist event bind type
Found by: Many
Patch by: Geo, thommey

A common issue for users writing scripts is trying to access the list of channel users as soon as Eggdrop joins (via JOIN bind), but Eggdrop has not yet received the list from the server. This event type will trigger every time Eggdrop finishes receiving the list of users for a channel (triggered by receiving a 315 message).

This also adds a new function check_tcl_event_arg which internally allows an optional argument to be passed, allowing event types to pass both the type of event and the optional argument to the event proc.