Properly queue/log QUIT message
Patch by: michaelortmann

Messages from eggdrop to irc server are logged when raw-log is set. depending on the queue the message was put in, it gets logged like for example [m->] or [s->] The Quit message was sent to server directly without using those queues and thus was not logged.